+++ src/mesido/esdl/profile_parser.py
@@ -38,6 +38,7 @@ class BaseProfileReader:
         "gas_demand": ".target_gas_demand",
         "gas_source": ".maximum_gas_source",
     }
+
     carrier_profile_var_name: str = ".price_profile"
 
     def __init__(self, energy_system: esdl.EnergySystem, file_path: Optional[Path]):
@@ -45,6 +46,11 @@ class BaseProfileReader:
         self._energy_system: esdl.EnergySystem = energy_system
         self._file_path: Optional[Path] = file_path
         self._reference_datetimes: Optional[pd.DatetimeIndex] = None
+        self._asset_potential_errors_identified: Dict[str, Dict] = {
+            "heat_demand.power": {},  # error type, heat demand name, error message
+            "cold_demand.power": {},  # error type, cold demand name, error message
+            "heat_demand.type": {},  # error type, heat demand name, error message
+        }
 
     def read_profiles(
         self,
@@ -54,7 +60,7 @@ class BaseProfileReader:
         esdl_assets: Dict[str, Asset],
         carrier_properties: Dict[str, Dict],
         ensemble_size: int,
-    ) -> None:
+    ) -> Dict[str, Dict]:
         """
         This function takes a datastore and a dictionary of milp network components and loads a
         profile for each demand and source in the provided milp network components into the
@@ -100,12 +106,6 @@ class BaseProfileReader:
             zip(esdl_asset_id_to_name_map.values(), esdl_asset_id_to_name_map.keys())
         )
 
-        # Define variable containing potential error information
-        asset_potential_errors = {
-            "heat_demand.power": {},  # error type, heat demand name, error message
-            "cold_demand.power": {},  # error type, cold demand name, error message
-        }
-
         for ensemble_member in range(ensemble_size):
             for component_type, var_name in self.component_type_to_var_name_map.items():
                 for component in energy_system_components.get(component_type, []):
@@ -135,7 +135,9 @@ class BaseProfileReader:
                     if component_type in ["heat_demand", "cold_demand"]:
                         max_profile_value = max(values)
                         if asset_power < max_profile_value:
-                            asset_potential_errors[f"{component_type}.power"][component] = (
+                            self._asset_potential_errors_identified[f"{component_type}.power"][
+                                component
+                            ] = (
                                 f"{component}: The installed capacity of"
                                 f" {round(asset_power / 1.0e6, 3)}MW should be larger than the"
                                 " maximum of the heat demand profile "
@@ -157,7 +159,7 @@ class BaseProfileReader:
                         values=profile,
                         ensemble_member=ensemble_member,
                     )
-        return asset_potential_errors
+        return self._asset_potential_errors_identified
 
     def _load_profiles_from_source(
         self,
@@ -193,6 +195,7 @@ class BaseProfileReader:
 class InfluxDBProfileReader(BaseProfileReader):
     asset_type_to_variable_name_conversion = {
         esdl.esdl.HeatingDemand: ".target_heat_demand",
+        esdl.esdl.GenericConsumer: ".target_heat_demand",
         esdl.esdl.HeatProducer: ".maximum_heat_source",
         esdl.esdl.ElectricityDemand: ".target_electricity_demand",
         esdl.esdl.ElectricityProducer: ".maximum_electricity_source",
@@ -295,12 +298,21 @@ class InfluxDBProfileReader(BaseProfileReader):
                 var_base_name = asset.name
                 try:
                     variable_suffix = self.asset_type_to_variable_name_conversion[type(asset)]
+                    # For multicommidity work profiles need to be assigned to GenericConsumer, but
+                    # not for heat network (this asset_potential_errors is used in grow_workflow)
+                    if type(asset) is esdl.GenericConsumer:
+                        # asset_potential_errors
+                        self._asset_potential_errors_identified["heat_demand.type"][asset.name] = (
+                            f"Asset {asset.name}: This asset is currently a GenericConsumer please"
+                            " change it to a HeatingDemand"
+                        )
                 except KeyError:
                     raise RuntimeError(
                         f"The asset {profile.field} is of type {type(asset)} which is "
                         f"currently not supported to have a profile to be loaded "
                         f"from the database."
                     )
+                    exit(1)
             else:
                 raise RuntimeError(
                     f"Got a profile for a {container}. Currently only profiles "