+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # [Unreleased] - 2024-08-20
 
 ## Added
+- User feedback regarding demand type to be used in heat networks 
 - Functionality to connect assets with logical links 
 
 ## Changed
+- Increased time limit for stage 1, priority 1
 - Replaced Retry loop to connect all assets and apply the correct modifiers of the linked assets
  
 ## Fixed
-- xxxx 
+- Error exit when stage 1, priority 1 was not successful
 
 # [0.1.7] - 2024-08-16